image-to-text-service/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
jinja2
